# designer-portfolio/server/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FolioStack API",
    "description": "API documentation for FolioStack - Portfolio Management Platform",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3002/api",
      "description": "Development server"
    },
    {
      "url": "https://api.foliostack.app/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/works": {
      "get": {
        "tags": ["Works"],
        "summary": "Get all works",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["draft", "published", "archived"] } },
          { "name": "category", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Works retrieved successfully" }
        }
      },
      "post": {
        "tags": ["Works"],
        "summary": "Create a new work",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "url": { "type": "string" },
                  "tags": { "type": "array", "items": { "type": "string" } },
                  "category": { "type": "string" },
                  "status": { "type": "string", "enum": ["draft", "published"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Work created successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/works/{id}": {
      "get": {
        "tags": ["Works"],
        "summary": "Get work by ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Work retrieved successfully" },
          "404": { "description": "Work not found" }
        }
      },
      "patch": {
        "tags": ["Works"],
        "summary": "Update work",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "url": { "type": "string" },
                  "tags": { "type": "array", "items": { "type": "string" } },
                  "category": { "type": "string" },
                  "status": { "type": "string", "enum": ["draft", "published", "archived"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Work updated successfully" },
          "404": { "description": "Work not found" }
        }
      },
      "delete": {
        "tags": ["Works"],
        "summary": "Delete work",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Work deleted successfully" },
          "404": { "description": "Work not found" }
        }
      }
    },
    "/works/import-url": {
      "post": {
        "tags": ["Works"],
        "summary": "Import work from URL",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "url": { "type": "string", "format": "uri" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Work imported successfully" },
          "400": { "description": "Invalid URL" },
          "409": { "description": "URL already imported" }
        }
      }
    },
    "/works/import-github": {
      "post": {
        "tags": ["Works"],
        "summary": "Import work from GitHub",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "url": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "GitHub repository imported successfully" },
          "400": { "description": "Invalid GitHub URL" },
          "404": { "description": "Repository not found" }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}